python-service/requirements.txt
fastapi
uvicorn
python-multipart
rembg[cpu]
python-dotenv
onnxruntime